pyproject.toml
[project]
name = "ai-research-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "gradio>=6.9.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "pydantic>=2.12.5",
    "pypdf>=6.8.0",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.3.0",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers>=5.3.0",
]
